### 2024-08-02 0ec0c71
modify Makefiles for reordered components
