### 2024-08-02 #251 @zdelrosario
(Moved `paper.md` into directory; testing if paper source is still found.)
