+++ .github/workflows/draft-pdf.yml
@@ -13,7 +13,7 @@ jobs:
         with:
           journal: joss
           # This should be the path to the paper within your repo.
-          paper-path: paper.md
+          paper-path: jose-paper/paper.md
       - name: Upload
         uses: actions/upload-artifact@v3
         with:
@@ -21,4 +21,4 @@ jobs:
           # This is the output path where Pandoc will write the compiled
           # PDF. Note, this should be the same directory as the input
           # paper.md
-          path: paper.pdf
+          path: jose-paper/paper.pdf
