### 2024-08-02 09d41ad
add link to CEJST page
