### 2024-08-02 f625fe6
link session info from README
