+++ README.md
@@ -6,6 +6,8 @@ A single-class activity that introduces the differences between frequentist and 
 - [Run-of-show document](https://github.com/bayes-bats/tier2-freq-bayes/blob/main/development/run-of-show.md)
 - Learning Objectives are detailed in the [01-introduction document](https://github.com/bayes-bats/tier2-freq-bayes/blob/main/development/01-introduction-main.qmd)
 
+Session info from a successful build of the activity materials are available in [this file](https://raw.githubusercontent.com/bayes-bats/tier2-freq-bayes/main/session-info).
+
 This activity was created by Stefani Langehennig and Zachary del Rosario as part of the [Bayes BATS](https://www.stat.uci.edu/bayes-bats/materials.html) program, with mentorship and advising from Mine Dogucu.
 
 This material is based upon work supported by the National Science Foundation under Grant No.s #2215879, #2215920, and #2215709.
