### 2024-08-02 a5982cf
link session info from JOSE paper draft
