### 2025-07-03 #251 @ethanwhite
Hi @zdelrosario  - Apologies for the long delay since your initial submission. Please let me know if you're still interested in having this paper reviewed at JOSE. If you are, I'll work on getting the review process started ASAP.
