### 2025-07-07 #251 @ethanwhite
@zdelrosario great! Can you take a look at [the list of people who have agreed to review for JOSE](https://bit.ly/jose-reviewers) and see if you have any suggestions for potential reviewers who may be suitable for this submission (recommendations not on this list are also welcome). Please provide their github usernames here in this thread, but without tagging them with an @. Once I have your list of I'll work on getting the review process started.
