### 2025-07-18 #294 @editorialbot
[REVIEW]: Productive Controversy: A Classroom Activity to Contrast Frequentist and Bayesian Paradigms
**Submitting author:** @zdelrosario (<a href="http://orcid.org/0000-0003-4676-1692">Zachary del Rosario</a>)
**Repository:** https://github.com/bayes-bats/tier2-freq-bayes
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
