### 2025-08-01 #294 @ethanwhite
👋 @rudeboybert, @valentina-s. Thanks again for agreeing to review this paper. Since we're now two weeks into the review process, I just wanted to check in and see if either of you have any questions or have run into any issues that I can help out with as you work on your reviews.
