### 2025-08-18 #294 @ethanwhite
👋 @rudeboybert, @valentina-s. We're now a little over 4 weeks into the review, which means you've got a little under two weeks to finish things up. Please let me know if either you have any questions or have run into any issues that I can help out with as you work on your reviews.
