### 2025-09-01 bayes-bats/tier2-freq-bayes#5 @rudeboybert
Add Documentation items in review checklist to repo's README
Hello, I'm currently [reviewing your article for JOSE](https://github.com/openjournals/jose-reviews/issues/294#issuecomment-3243197884). Could you add the 4 checklist items under the Documentation section to your repo's main README.md so that they are easy to find?
