### 2025-09-01 #294 @rudeboybert
https://github.com/bayes-bats/tier2-freq-bayes/issues/5
