### 2025-09-01 bayes-bats/tier2-freq-bayes#6 @rudeboybert
Create explicit GitHub repository release to match version
Hello, I'm currently [reviewing your article for JOSE](https://github.com/openjournals/jose-reviews/issues/294#issuecomment-3243197884). The version listed at the top is v1.0, but there is no explicit repository release matching this, as is required in the General Checks section of the review checklist. Could you tag a v1.0 release?
