### 2025-09-01 bayes-bats/tier2-freq-bayes#7 @rudeboybert
Add DOI
Hello, I'm currently [reviewing your article for JOSE](https://github.com/openjournals/jose-reviews/issues/294#issuecomment-3243197884). In the JOSE paper section of the review checklist, it asks for DOI's for all references. Could you double check the references in your [bib file](https://github.com/bayes-bats/tier2-freq-bayes/blob/main/jose-paper/paper.bib) that don't have DOI's? For example, I found this article's DOI https://www.researchgate.net/publication/318276150_Statistical_literacy_for_active_citizenship_A_call_for_data_science_education
