### 2025-09-02 bayes-bats/tier2-freq-bayes#8 @valentina-s
Mismatch between files referenced in paper vs repo
### [JOSE Review](https://github.com/openjournals/jose-reviews/issues/294#issuecomment-3238289735) (Documentation: Usage)

In section Contents of the paper, the listed sections do not match the filenames in the repo:
* `02a-activity-freq` vs `03a-activity-freq-main.qmd`
* `02b-activity-bayes` vs `03b-activity-bayes`-main.qmd`
* `03-simiplified` vs `02-one-pager-main`, but there is also `03-simplified-main`
