### 2025-09-02 bayes-bats/tier2-freq-bayes#9 @valentina-s
Improve installation/build documentation
### [JOSE Review](https://github.com/openjournals/jose-reviews/issues/294#issuecomment-3238289735) (Documentation: Installation Instructions)

1. It is not entirely obvious what the purpose of the Makefile is and what it is doing: provide more details in the repo Readme since there it will be most visible.
2. Upon trying to run `make` from the main repo, I encountered errors (possibly due to filename changes):<img width="519" height="113" alt="Image" src="https://github.com/user-attachments/assets/2a4fe283-616d-4e7e-8aac-46da840c48d6" />

3. How can a new teacher recreate the environment quickly? I see there is `session.info` but an actual environment file such as `renv.lock` and example how to recreate it will be very useful.
