### 2025-09-02 #294 @ethanwhite
Thanks to @rudeboybert and @valentina-s for making progress on their checklists and opening issues in core repo. Let me know when you've completed your initial reviews and then I'll look over the issues and check in with @zdelrosario
