### 2025-09-02 bayes-bats/tier2-freq-bayes#7 @rudeboybert
Maybe add `doi` fields for all references, and even if no doi exists leave them blank to indicate a good faith effort was made to find them?
