### 2025-09-02 bayes-bats/tier2-freq-bayes#10 @valentina-s
Audience experience clarification
### [JOSE Review](https://github.com/openjournals/jose-reviews/issues/294#issuecomment-3238289735) (Documentation: Statement of need)

Could you add a description of your target student audience in the paper and the prerequisites? The Readme mentions linear models and statistical inference.  Statistical inference is broad and can be taught at different levels. I wonder what the actual expectations are: for example, many undergraduates might have learnt about fitting lines but would not have heard about the notion of Maximum Likelihood, Prior, Posterior, and it is possible they may have not even learnt the notion of parameter estimation (they could have learnt testing in a Stat 101 class). You could add some notes about what the instructors need to explain in advance if the students are missing some prerequisites.
