### 2025-09-02 bayes-bats/tier2-freq-bayes#11 @valentina-s
Clarify usage of content in different file formats
### [JOSE Review](https://github.com/openjournals/jose-reviews/issues/294#issuecomment-3238289735) (Documentation: Usage)

Currently the content is in `qmd`, `pdf`, and `html`. Seems the activity was done in paper format, by giving them printouts of the `pdf`. Was it done in computational format: they need to execute the notebooks in R? What about the `html` format: I thought the intent of the `html`s was to stitch them in a website (on github pages or readthedocs similar to the [Carpentries lessons](https://carpentries.org/lessons/)) but I did not see one? I believe that can help new instructors and increase adoption.
