### 2025-09-02 bayes-bats/tier2-freq-bayes#12 @valentina-s
Community guidelines
### [JOSE Review](https://github.com/openjournals/jose-reviews/issues/294#issuecomment-3238289735) (Documentation: Community Guidelines)

Are there clear guidelines for third parties wishing to

1. Contribute to the module
2. Report issues or problems with the module
3. Seek support

Here is an example of [Contributing Guide](https://github.com/swcarpentry/python-novice-inflammation/blob/main/CONTRIBUTING.md) from the Carpentries. It does not need to be that elaborate as long as the above questions are answered, but this can serve as a reference on what to include.
