### 2025-09-02 #294 @valentina-s
@zdelrosario Thank you for your submission: I think this is a very useful activity that can be adopted by other researchers, and you have been very intentional about the learning objectives and have shared your experience running the activity. I have completed your review now and have added a few issues to the repo particularly on adding documentation details that can help new instructors. I also agree with the issues already added by @rudeboybert. Feel free to ask me if you have questions regarding my comments.
