### 2025-09-08 #294 @rudeboybert
Hey @ethanwhite sorry for the delay. This exercise strikes me as an excellent starting point to trying to bridge the gap that exists between conventional frequentist stats pedagogy and Bayesian thinking. I commend the authors for a very thorough job.
